requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
scipy>=1.11.0
anthropic>=0.7.0
google-generativeai>=0.3.0
openai>=1.0.0
yfinance>=0.2.28
pytz>=2023.3
python-dateutil>=2.8.2
gspread>=6.0.0
openpyxl>=3.1.0
